requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
reportlab>=4.0.0
pydantic>=2.0.0
holidays>=0.30.0
python-jose>=3.3.0
python-dateutil>=2.8.0
cachetools>=5.3.0
python-dotenv>=1.0.0
alembic>=1.13.0
requests>=2.31.0
python-json-logger>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
